combine comment, recipe

+++ fridge/src/main/java/exProject/fridge/repository/UserRecipeFavoriteRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface UserRecipeFavoriteRepository extends JpaRepository<UserRecipeFavorite, Long> {
+public interface UserRecipeFavoriteRepository extends JpaRepository<UserRecipeFavorite, Integer> {
 
     List<UserRecipeFavorite> findByUser(User user);
 }
